Better listin to myself this yime
I mentioned how horrible my relationships always turn out and seen myself as the common factor and thought myself as the problem. So before this one gets ruined and I think it won't at all but I will for the first time becfsithfull and quit leying and getting on hookup and dating sites. Meet-seniors.com is like throwing it in her face. And she knows. She always knows